fixup! Add setting for timeout-reboot

+++ src/com/android/settings/security/DeviceRebootTimeoutPreferenceController.java
@@ -65,8 +65,8 @@ public class DeviceRebootTimeoutPreferenceController extends BasePreferenceContr
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
         try {
-            int value = Integer.parseInt((String) newValue);
-            LineageSettings.Global.putInt(mContext.getContentResolver(),
+            long value = Long.parseLong((String) newValue);
+            LineageSettings.Global.putLong(mContext.getContentResolver(),
                     LineageSettings.Global.DEVICE_REBOOT_TIMEOUT, value);
             updateTimeoutPreferenceDescription((ListPreference) preference, value);
         } catch (NumberFormatException e) {
